chore: move notifications sms pull test to large machine
