+++ packages/amplify-cli-core/src/errors/amplify-exception.ts
@@ -181,4 +181,5 @@ export type AmplifyFaultType =
   | 'UnexpectedS3Fault'
   | 'UnknownFault'
   | 'UnknownNodeJSFault'
-  | 'MockProcessFault';
+  | 'MockProcessFault'
+  | 'ZipExtractFault';
