+++ scripts/split-e2e-tests.ts
@@ -391,7 +391,7 @@ function validateArtifactStoragePaths(config: CircleCIConfig) {
   for(let key of Object.keys(config.jobs)) {
     const job = config.jobs[key];
     const steps = job.steps;
-    
+
     for(let i = 0; i < steps.length; i ++){
       const resultsPath = steps[i].store_test_results;
       const artifactsPath = steps[i].store_artifacts;
@@ -436,7 +436,7 @@ function validateArtifactStoragePaths(config: CircleCIConfig) {
 
 function main(): void {
   const config = loadConfig();
-  
+
   validateArtifactStoragePaths(config);
 
   const splitPkgTests = splitTests(
